pyproject.toml
[project]
name = "price-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
